+++ CHANGES.rst
@@ -3,6 +3,11 @@ hmmlearn Changelog
 
 Here you can see the full list of changes between each hmmlearn release.
 
+next
+----
+
+- MultinomialHMM now accepts unsigned symbols (#258).
+
 Version 0.2.1
 -------------